# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Nava Skills Hub API
  version: 1.0.0
  description: |
    REST API for the Nava Skills Hub. All endpoints require a valid `__session` cookie
    (HS256 JWT issued by the auth flow). Obtain one by completing the Google OAuth flow
    at `/auth/login`.

    **Base URLs:**
    - Production: `https://hub.navapbc.com`
    - Staging: `https://staging.hub.navapbc.com`

    **Roles:** the ladder is `user` (default) → `maintain` → `admin`, where each role inherits
    all permissions from the roles below it. `projects-admin` is a capability role that sits
    *outside* that ladder: it grants management of Contract Explorer reference data and nothing
    else privileged. See [rbac-permissions.md](rbac-permissions.md) for the full permission matrix.
servers:
  - url: https://hub.navapbc.com
    description: Production
  - url: https://staging.hub.navapbc.com
    description: Staging
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __session
      description: HS256 JWT issued after Google OAuth. 8-hour expiry.
  schemas:
...
    ProjectReferenceRecord:
      type: object
      description: >
        A Contract Explorer reference record. Archetypes and postures share one
        table and one shape; `icon`, `description`, `characteristics` and
        `ai_opportunities` apply to archetypes, `position` and `steps` to postures.
      required: [id, label, color]
      properties:
        entity_type:
          type: string
          enum: [archetype, posture]
          readOnly: true
        id:
          type: string
          description: >
            Slug-safe — lowercase alphanumeric words separated by hyphens. Doubles
            as a DOM and style key in the Contract Explorer.
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          example: product-team
        label:
          type: string
          example: Product Team
        color:
          type: string
          description: >
            Six-digit hex, applied as an inline style. For a posture this is the
            badge background rather than a hue to blend, so it pairs with a fixed
            dark foreground.
          pattern: "^#[0-9a-fA-F]{6}$"
          example: "#651A94"
        status:
          type: string
          enum: [active, inactive]
          default: active
          description: Set via the status endpoint, not by update. There is no delete.
        description:
          type: string
          description: Archetypes only.
        icon:
          type: string
          description: >
            Archetypes only. Must be a name from the curated menu in
            src/lib/icons.mjs — an unlisted name would render as nothing.
          example: users
        characteristics:
          type: array
          description: Archetypes only. Order is preserved.
          items:
            type: string
        ai_opportunities:
          type: array
          description: Archetypes only. Order is preserved.
          items:
            type: string
        position:
          type: integer
          description: >
            Postures only. Display order wherever postures are enumerated. Carries
            no severity or priority meaning — nothing branches on it.
        steps:
          type: array
          description: >
            Postures only. Ordered guidance steps; order is semantically
            load-bearing and preserved as authored. No entry may be empty.
          items:
            type: string
        created_by:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_by:
          type: string
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Forbidden
security:
  - sessionCookie: []
paths:
...
  /api/project-reference/{entityType}:
    parameters:
      - name: entityType
        in: path
        required: true
        description: Which reference dataset. Anything else is a 400.
        schema:
          type: string
          enum: [archetype, posture]

    get:
      tags: [project-reference]
      summary: List reference records of one entity type
      description: >
        Requires `projects-admin` or `admin`. Unlike the skills and plugins list
        endpoints, this read is authorised — the data is admin-owned. Returns
        active and inactive records alike.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProjectReferenceRecord"
        "400":
          description: Unknown entity type
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags: [project-reference]
      summary: Create a reference record
      description: Requires `projects-admin` or `admin`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectReferenceRecord"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectReferenceRecord"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: A record with this id already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/project-reference/{entityType}/{id}:
    parameters:
      - name: entityType
        in: path
        required: true
        schema:
          type: string
          enum: [archetype, posture]
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      tags: [project-reference]
      summary: Get a reference record
      description: Requires `projects-admin` or `admin`. Deactivated records remain retrievable.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectReferenceRecord"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags: [project-reference]
      summary: Replace a reference record
      description: >
        Requires `projects-admin` or `admin`. The path owns the id; an `id` in the
        body is ignored, so an edit cannot move a record. Whole-record write —
        last writer wins, which is acceptable for this low-churn data.
        There is no delete endpoint: a referenced record must not be removable,
        so deactivate it via the status endpoint instead.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectReferenceRecord"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectReferenceRecord"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/project-reference/{entityType}/{id}/status:
    parameters:
      - name: entityType
        in: path
        required: true
        schema:
          type: string
          enum: [archetype, posture]
      - name: id
        in: path
        required: true
        schema:
          type: string

    put:
      tags: [project-reference]
      summary: Activate or deactivate a reference record
      description: >
        Requires `projects-admin` or `admin`. Separate from the update endpoint so
        the audit trail records `deactivated` / `reactivated` rather than a
        generic update. Deactivating stops a record being offered as a choice
        while existing references to it continue to resolve.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [active, inactive]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectReferenceRecord"
        "400":
          description: Invalid status value
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"